add `ui/time` tests
